src/requirements.txt
requests
uiautomation
loguru
python-dotenv
flet
certifi
